ai-service-python/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.2
typing_extensions==4.15.0
